Refactor: NotionPage 파일 삭제
메인 페이지가 notionPage이고, 다른 페이지가 존재하지 않으므로 App.js에 모두 구현하도록 변경
