+++ src/components/App.js
@@ -1,16 +1,17 @@
-import { getRootDocument, getSpecificDocument } from '../api/api'
-import NotionPage from '../pages/NotionPage'
+import { createDocument, deleteSpecificDocument, getRootDocument, getSpecificDocument } from '../api/api'
+import Content from '../components/Content'
+import DocumentsList from '../components/DocumentsList'
 
 export default function App({ $target, initialState }) {
-  const $app = document.createElement('div')
-  $target.appendChild($app)
+  const $notionPage = document.createElement('div')
+  $target.appendChild($notionPage)
 
-  // documents state
+  // documents list state
   this.documents = initialState.documents
 
   this.setDocuments = (nextDocuments) => {
     this.documents = nextDocuments
-    notionPage.documentsList.setState(nextDocuments)
+    documentsList.setState(nextDocuments)
   }
 
   // document editor state
@@ -18,8 +19,37 @@ export default function App({ $target, initialState }) {
 
   this.setEditorContent = (nextContent) => {
     this.editorContent = nextContent
-    notionPage.content.setState(nextContent)
+    content.setState(nextContent)
   }
+
+  const documentsList = new DocumentsList({ 
+    $target: $notionPage, 
+    initialState: this.documents,
+    onClick: (id) => {
+      fetchContentData(id)
+    },
+    onAdd: async (id) => {
+      const data = {
+        title: "제목 없음",
+        parent: id ? id : null
+      }
+
+      const newDocument = await createDocument(data)
+
+      fetchDocumentsData()
+      fetchContentData(newDocument.id)
+    },
+  })
+
+  const content = new Content({
+    $target: $notionPage,
+    initialState: this.editorContent,
+    onDelete: async (id) => {
+      await deleteSpecificDocument(id)
+      fetchDocumentsData()
+      fetchContentData()
+    }
+  })
 
   // 함수 파일 작성하기
   // 전체 documentsList 불러오기
@@ -37,13 +67,6 @@ export default function App({ $target, initialState }) {
       this.setEditorContent(null)
     }
   }
-
-  const notionPage = new NotionPage({
-    $target,
-    initialState,
-    fetchContentData,
-    fetchDocumentsData,
-  })
 
   fetchDocumentsData()
 }
